+++ lib/util/extensions/enemy/rotation_enemy_extensions.dart
@@ -11,7 +11,8 @@ extension RotationEnemyExtensions on RotationEnemy {
   /// Checks whether the player is within range. If so, move to it.
   void seeAndMoveToPlayer({
     required Function(Player) closePlayer,
-    VoidCallback? notObserved,
+    // return true to stop move.
+    BoolCallback? notObserved,
     double radiusVision = 32,
     double margin = 10,
     bool runOnlyVisibleInScreen = true,
@@ -37,9 +38,8 @@ extension RotationEnemyExtensions on RotationEnemy {
         moveFromAngle(radAngle);
       },
       notObserved: () {
-        if (notObserved != null) {
-          notObserved();
-        } else {
+        bool stop = notObserved?.call() ?? true;
+        if (stop) {
           stopMove();
         }
       },
@@ -49,7 +49,8 @@ extension RotationEnemyExtensions on RotationEnemy {
   /// Checks whether the player is within range. If so, move to it.
   void seeAndMoveToAttackRange({
     required Function(Player) positioned,
-    VoidCallback? notObserved,
+    // return true to stop move.
+    BoolCallback? notObserved,
     double radiusVision = 32,
     double? minDistanceCellsFromPlayer,
     bool runOnlyVisibleInScreen = true,
@@ -87,9 +88,8 @@ extension RotationEnemyExtensions on RotationEnemy {
         );
       },
       notObserved: () {
-        if (notObserved != null) {
-          notObserved();
-        } else {
+        bool stop = notObserved?.call() ?? true;
+        if (stop) {
           stopMove();
         }
       },